simplest prototype: MatlabControlTask works as a replacement for MatlabTask, and MatlabControlEngine deals with the communication with MatlabControl
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/scheduling/trunk@19837 28e8926c-6b08-0410-baaa-805c5e19b8d6

+++ src/scheduler/src/org/ow2/proactive/scheduler/ext/matlab/worker/MatlabTask.java
@@ -280,24 +280,26 @@ public class MatlabTask<W extends AOMatlabWorker> extends
                 "] Killing selection process...");
         }
         Process p = matlabEngineConfig.getSelectionScriptProcess();
-        if (os.equals(OperatingSystem.windows)) {
-            WinProcess pi = new WinProcess(p);
-            try {
-                if (paconfig.isDebug()) {
-                    System.out.println("Killing process " + pi.getPid());
-                    outDebug.println("Killing process " + pi.getPid());
-                }
-                Runtime.getRuntime().exec("taskkill /PID " + pi.getPid() + " /T");
-                Runtime.getRuntime().exec("tskill " + pi.getPid());
-                killProcessWindowsWithEnv("SELECTION_SCRIPT", nodeName);
+        if (p != null) {
+            if (os.equals(OperatingSystem.windows)) {
+                WinProcess pi = new WinProcess(p);
+                try {
+                    if (paconfig.isDebug()) {
+                        System.out.println("Killing process " + pi.getPid());
+                        outDebug.println("Killing process " + pi.getPid());
+                    }
+                    Runtime.getRuntime().exec("taskkill /PID " + pi.getPid() + " /T");
+                    Runtime.getRuntime().exec("tskill " + pi.getPid());
+                    killProcessWindowsWithEnv("SELECTION_SCRIPT", nodeName);
 
-            } catch (IOException e) {
-                e.printStackTrace();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            } else {
+                p.destroy();
             }
-        } else {
-            p.destroy();
         }
-        matlabEngineConfig.setSelectionScriptProcess(p);
+        matlabEngineConfig.setSelectionScriptProcess(null);
     }
 
     protected void initWorker(W worker) throws Throwable {
